requirements.txt
streamlit==1.12.2
pandas
snowflake-connector-python
toml
altair
psycopg2-binary==2.9.3
